Setting hardware for PIC18F2550 is ok

+++ flash.h
@@ -4,7 +4,7 @@
 #include "GenericTypeDefs.h"
 #include "Compiler.h"
 
-void writeEEPROM(UINT16, UINT16);
-UINT16 readEEPROM(UINT16);
+void writeEEPROM(UINT8, UINT16);
+UINT16 readEEPROM(UINT8);
 
 #endif	/* FLASH_H */
